fix: side menu not opening
